requirements.txt
fastapi
uvicorn
pydantic
sse-starlette
python-multipart
pymata4==1.15
pyserial
